move config so not exported
