+++ abn/config.go
@@ -1,4 +1,4 @@
-package watcher
+package abn
 
 // config.go - reading of configurtion (list of resources/namespaces to watch)
 
@@ -11,18 +11,18 @@ import (
 	"sigs.k8s.io/yaml"
 )
 
-// Config is A/B(/n) serivce configuration. The configuration identifies the resources to watch and in which namespaces.
-type Config struct {
+// config is A/B(/n) serivce configuration. The configuration identifies the resources to watch and in which namespaces.
+type config struct {
 	// Namespaces is list of namespaces to watch
 	Namespaces []string `yaml:"namespaces,omitemtpy"`
 	// Resources is list of resoure types that should be watched in the namespaces
 	Resources []schema.GroupVersionResource `yaml:"resources,omitemtpy"`
 }
 
-// ReadConfig reads yaml config file fn and converts to a Config object
-func ReadConfig(fn string) (config Config) {
+// readConfig reads yaml config file fn and converts to a Config object
+func readConfig(fn string) (c config) {
 	// empty configuration
-	config = Config{
+	c = config{
 		Namespaces: []string{},
 		Resources:  []schema.GroupVersionResource{},
 	}
@@ -30,23 +30,23 @@ func ReadConfig(fn string) (config Config) {
 	yfile, err := ioutil.ReadFile(fn)
 	if err != nil {
 		log.Logger.Warnf("unable to read configuration file %s: %s", fn, err.Error())
-		return config // empty configuration
+		return c // empty configuration
 	}
 
 	log.Logger.Debugf("read configuration\n%s", string(yfile))
 
-	err = yaml.Unmarshal(yfile, &config)
+	err = yaml.Unmarshal(yfile, &c)
 	if err != nil {
 		log.Logger.Warnf("invalid configuration file %s: %s", fn, err.Error())
-		return config // empty configuration
+		return c // empty configuration
 	}
 
-	if len(config.Namespaces) == 0 {
+	if len(c.Namespaces) == 0 {
 		log.Logger.Warn("not watching any namespaces - configuration error?")
 	}
-	if len(config.Resources) == 0 {
+	if len(c.Resources) == 0 {
 		log.Logger.Warn("not watching any resources - configuration error?")
 	}
 
-	return config
+	return c
 }
